+++ app/src/androidTest/java/woowacourse/payments/ui/payments/registration/CardholderNameTextFieldTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.payments.registration
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -15,7 +15,6 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.payments.ui.model.CardholderNameUiModel
-import woowacourse.payments.ui.payments.CardholderNameTextField
 
 class CardholderNameTextFieldTest {
     @get:Rule
