+++ app/src/main/java/woowacourse/payments/ui/payments/registration/PaymentCardRegistrationActivity.kt
@@ -1,13 +1,12 @@
-package woowacourse.payments
+package woowacourse.payments.ui.payments.registration
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import woowacourse.payments.ui.payments.CardRegistrationScreen
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
-class MainActivity : ComponentActivity() {
+class PaymentCardRegistrationActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
